feat(refactor): refactor property tests
